combine deletion modal
